+++ elastic4s-core/src/main/scala/com/sksamuel/elastic4s/requests/searches/sort/ScriptSort.scala
@@ -1,6 +1,5 @@
 package com.sksamuel.elastic4s.requests.searches.sort
 
-import com.sksamuel.elastic4s.api.ScriptSortType
 import com.sksamuel.elastic4s.requests.script.Script
 import com.sksamuel.elastic4s.requests.searches.queries.Query
 import com.sksamuel.exts.OptionImplicits._
