+++ elastic4s-domain/src/main/scala/com/sksamuel/elastic4s/requests/searches/suggestion/CompletionSuggestion.scala
